fix(rave-mark): ensure mail label prints at correct size (#21)
Chrome behaves unexpectedly when printing multiple back-to-back documents that have different `@page` sizes. Either the ballot page gets scrunched to 4x6 or the label gets printed to a letter-size canvas. This fixes the issue by printing to PDF in the backend using puppeteer. It's a little heavy-handed, but it works pretty well.