various improvements to JX/Mark interop (#96)
* various improvements to JX/Mark interop

The serialization scheme got out of sync between the two sides. This brings them back in line. Also ensures that the card used to authenticate with JX terminal is the expected one. This will change once we use the TPM.

* test(mark): ensure signed objects have valid signatures

Now that we verify the signature before we add to the data store, we need the signatures to be real.