fix(jx): improve reliability of smartcard detection and status stream (#97)
Replaces `try_stream!` with a broadcast stream. This means that no matter how many requests we get for the status stream endpoint we only collect the data for it in one place. Previously, the loop inside the `try_stream!` would run even after the connection was closed, resulting in excess resource usage.