feat(jx): verify card PIN on insert (#162)
* feat(jx): verify card PIN on insert

This changes a bunch of stuff in order to fix an issue with the interaction between sync & async code, standardizing on async code with tokio. Various parts of the code have to either bridge with sync code such as `pcsc::Card` or bridge threads while holding data that isn't `Send` or `Sync`, such as handles to `pcsc::Card`. As a result, a lot of channels are used to manage this bridging.

* chore: remove unused `pin-project` dependency