feat(mark): allow skipping registration in usability tests (#186)
When running a usability test, run with `USABILITY_TEST_SKIP_REGISTRATION=true` to automatically skip the registration flow and go straight into voting. Note that the `RegistrationRequest` object will not be signed by the CAC, but this shouldn't matter as we're not actually syncing with any service that will attempt to verify that.